[feat]: 산책 경로 생성 구현

See merge request s09-bigdata-dist-sub2/S09P22B304!20

+++ backend/src/main/java/io/ssafy/p/j9b304/backend/domain/walk/dto/response/WalkGetDetailResponseDto.java
@@ -27,7 +27,7 @@ public class WalkGetDetailResponseDto {
     Integer time;
 
     @Schema(description = "이동 거리(km)")
-    Float distance;
+    Double distance;
 
     @Schema(description = "걸음 수")
     Integer walkCount;
